+++ AZ3166/AZ3166-1.0.0/libraries/AzureIotHub/examples/ShakeShake/ShakeShake.ino
@@ -261,7 +261,7 @@ static void DoIdle()
 
 static void DoShake()
 {
-  uint16_t steps = 0;
+  int steps = 0;
   acc_gyro->get_step_counter(&steps);
   if (steps > 2)
   {
